First working native macOS Captive Web View and better view handling for iOS and Catalyst
Fixed the view handling in the iOS library to be based on anchor constraints,
instead of setting the ViewController.view property. That resolved a UI glitch
in Catalyst. The default is now to constrain to the safe area.

Added a macOS framework to the Captive Web View library project, and a Mac
Skeleton application. The macOS version doesn't have all the programmatic UI
building, so applications must have a storyboard. The WKWebView can still be
created programmatically, and made to fill the view.

The framework for macOS doesn't have any substantial code of its own. It shares
some of the iOS code. The URLSchemeTaskError class was moved to facilitate that.

Moved the Skeleton web resources out from the Android assets to the repository
Web Resources directory.

The Captivity application has an input field that should get focus by default.
It now gets focus when the user selects the first tab again after selecting
another tab.

Focussing doesn't work on iOS at application load time. It works the rest of the
time.